demo_test.py
"""
Demo test for TalentScout Chatbot
"""
import time
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC

print("Starting automated demo test...")

# Note: This would require selenium and Chrome driver
# For now, just print manual instructions
print("\n" + "="*50)
print("MANUAL DEMO TEST INSTRUCTIONS")
print("="*50)
print("\n1. Open browser to: http://13.203.206.156:8501")
print("2. Verify the page loads with 'TalentScout AI Hiring Assistant'")
print("3. Check that chatbot greets you")
print("4. Test conversation flow:")
print("   - Enter: John Doe")
print("   - Enter: john@example.com")
print("   - Enter: 1234567890")
print("   - Enter: 5")
print("   - Enter: Software Engineer")
print("   - Enter: San Francisco")
print("   - Enter: Python, Django, PostgreSQL, React")
print("5. Verify OpenAI generates technical questions")
print("6. Type 'bye' to test graceful exit")
print("\nExpected Results:")
print("✅ All 7 information fields collected")
print("✅ OpenAI generates 3-5 questions per technology")
print("✅ Conversation context maintained")
print("✅ Graceful exit with 'bye'")
print("✅ Professional UI with custom styling")
